Added cofirmation dialog for delete actions, some component styling and few refactors

+++ src/app/projects/components/project-list/project-list.component.ts
@@ -1,11 +1,13 @@
-import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import {Component, OnInit} from '@angular/core';
+import {Router} from '@angular/router';
+import {MatSnackBar} from '@angular/material/snack-bar';
+import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material/dialog';
 
-import { remove } from 'lodash';
+import {remove} from 'lodash';
 
-import { Project } from './../../models/project';
-import { ProjectService } from './../../services/project.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import {Project} from '../../models/project';
+import {ProjectService} from '../../services/project.service';
+import {ConfirmDelDialogComponent} from './confirm-del-dialog.component';
 
 
 @Component({
@@ -17,11 +19,13 @@ export class ProjectListComponent implements OnInit {
 
   public displayedColumns: string[] = ['name', 'category', 'created', 'updated', 'action'];
   public dataSource: Project[] = [];
+  private confirmDelete = '';
 
   constructor(
     private projectService: ProjectService,
     private router: Router,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    public dialog: MatDialog
   ) { }
 
   ngOnInit(): void {
@@ -29,11 +33,10 @@ export class ProjectListComponent implements OnInit {
     .getProjets()
     .subscribe(data => {
       this.dataSource = data;
-    },err => this.handleError(err, 'Whoops, failed to fetch project list!'))
+    }, err => this.handleError(err, 'Whoops, failed to fetch project list!'));
   }
 
   private handleError(err: any, message: string): void {
-    console.log(err);
     this.snackBar.open(
       message, 'Error', {
         duration: 4000
@@ -42,7 +45,7 @@ export class ProjectListComponent implements OnInit {
   }
 
   createProjectActionHandler() {
-    this.router.navigate(['xyz', 'projects','new']);
+    this.router.navigate(['xyz', 'projects', 'new']);
   }
 
   updateProjectActionHandler(id: string) {
@@ -50,18 +53,33 @@ export class ProjectListComponent implements OnInit {
   }
 
   deleteProjectActionHandler(id: string) {
-    console.log(id);
-    this.projectService.deleteProject(id)
-    .subscribe(data => {
-      const removedItem = remove(this.dataSource, (item) => {
-        return item._id === data._id;
-      });
-      this.dataSource = [...this.dataSource]; /** Tells Angular to update variable. */
-      this.snackBar.open('Project deleted!', 'Success', {
-        duration: 8000
-      });
-    },err => this.handleError(err, 'Whoops, failed to delete project!')
+    const confirmationDialog = this.dialog.open(ConfirmDelDialogComponent, {
+        width: '350px',
+      }
     );
+    confirmationDialog.afterClosed().subscribe(result => {
+      // @todo Improve this check... Someone can hack its html attrib to pass this check
+      if (typeof result === 'string' && result === 'true') {
+        // User has confirmed to delete this item
+        this.projectService.deleteProject(id)
+          .subscribe(data => {
+              const removedItem = remove(this.dataSource, (item) => {
+                return item._id === data._id;
+              });
+              this.dataSource = [...this.dataSource]; /** Tells Angular to update variable. */
+              this.snackBar.open('Project deleted!', 'Success', {
+                duration: 8000
+              });
+            }, err => this.handleError(err, 'Whoops, failed to delete project!')
+          );
+      }
+    });
+
   }
 
+  private confirmDeleteDialog(): void {
+
+  }
 }
+
+
